Added a popup window for the image file to open, and some warnings, and some code cleanup.
