+++ msk2bmpGUI_C_C++_imgui_sdl2_tinyfiledialogs/imgui.ini
@@ -4,12 +4,22 @@ Size=400,400
 Collapsed=0
 
 [Window][Hello, world!]
-Pos=27,6
-Size=650,624
+Pos=32,39
+Size=587,515
 Collapsed=0
 
 [Window][Another Window]
-Pos=575,322
-Size=198,71
+Pos=126,211
+Size=229,142
+Collapsed=0
+
+[Window][Nothing selected...]
+Pos=552,278
+Size=311,196
+Collapsed=0
+
+[Window][Image Window...]
+Pos=449,69
+Size=739,549
 Collapsed=0
